Move json related tests to own file
